move primary generator action to G4user
it is a Geant4 user action, so it should be included in that submodule
